Add files via upload
A process that allows users to regain access to their accounts by changing their forgotten password. It's usually done by clicking a "Forgot password?" link or button on an authentication page. The system then generates a unique link that's sent to the user's email address. The link takes the user to a page where they can enter a new password.